결제 금액 타입 변경 (Long -> BigDecimal)

+++ src/main/java/com/xladmt/makify/challenge/dto/ChallengeCreateRequest.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
@@ -30,8 +31,8 @@ public class ChallengeCreateRequest {
     public String privateCode;
 
     public YN isFixedDeposit;
-    public Integer fixedDeposit;
-    public Integer maxDeposit;
+    public BigDecimal fixedDeposit;
+    public BigDecimal maxDeposit;
 
     public Category category;
     private List<MultipartFile> images;
